Profile: Save to the profile that was loaded at startup instead of "xcsoar-registry.prf"
This is much more intuitive behavior for something called "Profile"
Fixes ticket #470

Details:
This removes the concept of incremental profiles in favor of real profiles, which is a concept that is easier to
understand and more intuitive. Before this patch a user that loaded ProfileA.prf would wonder why changes
are not saved to that exact profile, but to xcsoar-registry.prf. If ProfileA.prf was loaded at next startup some
of the changes may have been reverted because they were not saved to ProfileA.prf. After this patch the profile
that is loaded at startup is also the one that will be saved to when changing settings. Instead of loading
xcsoar-registry.prf first and patching the loaded settings once again with the selected profile on startup,
now the hardcoded defaults are patched directly with the contents of e.g. ProfileA.prf.